backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pyjwt>=2.8.0
bcrypt>=4.0.0
requests>=2.31.0
python-multipart>=0.0.9
